requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
decorator==4.4.2
Django==3.1.1
djangorestframework==3.11.1
docutils==0.15.2
gensim==3.8.3
google-api-core==1.22.1
google-api-python-client==1.8.2
google-auth==1.21.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-speech==1.3.2
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.0
jmespath==0.10.0
moviepy==1.0.3
numexpr==2.7.1
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.1.1
sqlparse==0.3.1
tqdm==4.48.2
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0

pyLDAvis~=2.1.2